requirements.txt
certifi==2021.5.30
charset-normalizer==2.0.4
click==8.0.1
dnspython==2.1.0
Flask==2.0.1
Flask-RESTful==0.3.9
idna==3.2
isort==5.9.3
itsdangerous==2.0.1
Jinja2==3.0.1
lazy-object-proxy==1.6.0
Markdown==3.3.4
MarkupSafe==2.0.1
mccabe==0.6.1
numpy==1.21.2
opencv-python==4.5.3.56
platformdirs==2.3.0
pylint==2.10.2
pytz==2021.1
requests==2.26.0
six==1.16.0
toml==0.10.2
urllib3==1.26.6
Werkzeug==2.0.1
wrapt==1.12.1
python-dotenv
flask-cors